requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
tmdbsimple==2.9.1
djangorestframework==3.13.1
configparser==5.2.0
django-cache-memoize==0.1.10